refactor models to separate the learning process
